+++ src/main/java/io/cavia/trader/module/auth/controller/AuthRestController.java
@@ -3,10 +3,10 @@ package io.cavia.trader.module.auth.controller;
 import io.cavia.trader.module.auth.dto.LoginRequestDto;
 import io.cavia.trader.module.auth.dto.ResetPasswordRequestDto;
 import io.cavia.trader.module.auth.dto.SignupForm;
-import io.cavia.trader.module.auth.entity.Member;
 import io.cavia.trader.module.auth.security.UserDetailsImpl;
 import io.cavia.trader.module.auth.service.AuthService;
 import io.cavia.trader.module.jwt.JwtUtil;
+import io.cavia.trader.module.member.entity.Member;
 import jakarta.servlet.http.HttpServletResponse;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
