+++ src/main/java/io/cavia/trader/module/auth/service/AuthServiceImpl.java
@@ -6,9 +6,9 @@ import io.cavia.trader.module.auth.dto.LoginRequestDto;
 import io.cavia.trader.module.auth.dto.ResetPasswordRequestDto;
 import io.cavia.trader.module.auth.dto.SignupForm;
 import io.cavia.trader.module.auth.entity.EmailVerification;
-import io.cavia.trader.module.auth.entity.Member;
+import io.cavia.trader.module.member.entity.Member;
 import io.cavia.trader.module.auth.repository.EmailVerificationRepository;
-import io.cavia.trader.module.auth.repository.MemberRepository;
+import io.cavia.trader.module.member.repository.MemberRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.security.crypto.password.PasswordEncoder;
